the great sprite update

+++ src/data/pokemon/item_effects.h
@@ -1,84 +1,84 @@
-const u8 gItemEffect_Potion[7] = {
+const u16 gItemEffect_Potion[7] = {
     [4] = ITEM4_HEAL_HP,
     [6] = 20, // Amount of HP to recover
 };
 
-const u8 gItemEffect_Antidote[6] = {
+const u16 gItemEffect_Antidote[6] = {
     [3] = ITEM3_POISON,
 };
 
-const u8 gItemEffect_BurnHeal[6] = {
+const u16 gItemEffect_BurnHeal[6] = {
     [3] = ITEM3_BURN,
 };
 
-const u8 gItemEffect_IceHeal[6] = {
+const u16 gItemEffect_IceHeal[6] = {
     [3] = ITEM3_FREEZE,
 };
 
-const u8 gItemEffect_Awakening[6] = {
+const u16 gItemEffect_Awakening[6] = {
     [3] = ITEM3_SLEEP,
 };
 
-const u8 gItemEffect_ParalyzeHeal[6] = {
+const u16 gItemEffect_ParalyzeHeal[6] = {
     [3] = ITEM3_PARALYSIS,
 };
 
-const u8 gItemEffect_FullRestore[7] = {
+const u16 gItemEffect_FullRestore[7] = {
     [3] = ITEM3_STATUS_ALL,
     [4] = ITEM4_HEAL_HP,
     [6] = ITEM6_HEAL_HP_FULL,
 };
 
-const u8 gItemEffect_MaxPotion[7] = {
+const u16 gItemEffect_MaxPotion[7] = {
     [4] = ITEM4_HEAL_HP,
     [6] = ITEM6_HEAL_HP_FULL,
 };
 
-const u8 gItemEffect_HyperPotion[7] = {
+const u16 gItemEffect_HyperPotion[7] = {
     [4] = ITEM4_HEAL_HP,
     [6] = 200, // Amount of HP to recover
 };
 
-const u8 gItemEffect_SuperPotion[7] = {
+const u16 gItemEffect_SuperPotion[7] = {
     [4] = ITEM4_HEAL_HP,
     [6] = 50, // Amount of HP to recover
 };
 
-const u8 gItemEffect_FullHeal[6] = {
+const u16 gItemEffect_FullHeal[6] = {
     [3] = ITEM3_STATUS_ALL,
 };
 
-const u8 gItemEffect_Revive[7] = {
+const u16 gItemEffect_Revive[7] = {
     [4] = ITEM4_REVIVE | ITEM4_HEAL_HP,
     [6] = ITEM6_HEAL_HP_HALF,
 };
 
-const u8 gItemEffect_MaxRevive[7] = {
+const u16 gItemEffect_MaxRevive[7] = {
     [4] = ITEM4_REVIVE | ITEM4_HEAL_HP,
     [6] = ITEM6_HEAL_HP_FULL,
 };
 
-const u8 gItemEffect_FreshWater[7] = {
+const u16 gItemEffect_FreshWater[7] = {
     [4] = ITEM4_HEAL_HP,
     [6] = 50, // Amount of HP to recover
 };
 
-const u8 gItemEffect_SodaPop[7] = {
+const u16 gItemEffect_SodaPop[7] = {
     [4] = ITEM4_HEAL_HP,
     [6] = 60, // Amount of HP to recover
 };
 
-const u8 gItemEffect_Lemonade[7] = {
+const u16 gItemEffect_Lemonade[7] = {
     [4] = ITEM4_HEAL_HP,
     [6] = 80, // Amount of HP to recover
 };
 
-const u8 gItemEffect_MoomooMilk[7] = {
+const u16 gItemEffect_MoomooMilk[7] = {
     [4] = ITEM4_HEAL_HP,
     [6] = 100, // Amount of HP to recover
 };
 
-const u8 gItemEffect_EnergyPowder[10] = {
+const u16 gItemEffect_EnergyPowder[10] = {
     [4] = ITEM4_HEAL_HP,
     [5] = ITEM5_FRIENDSHIP_ALL,
     [6] = 50,  // Amount of HP to recover
@@ -87,7 +87,7 @@ const u8 gItemEffect_EnergyPowder[10] = {
     [9] = -10, // Friendship change, high
 };
 
-const u8 gItemEffect_EnergyRoot[10] = {
+const u16 gItemEffect_EnergyRoot[10] = {
     [4] = ITEM4_HEAL_HP,
     [5] = ITEM5_FRIENDSHIP_ALL,
     [6] = 200, // Amount of HP to recover
@@ -96,7 +96,7 @@ const u8 gItemEffect_EnergyRoot[10] = {
     [9] = -15, // Friendship change, high
 };
 
-const u8 gItemEffect_HealPowder[9] = {
+const u16 gItemEffect_HealPowder[9] = {
     [3] = ITEM3_STATUS_ALL,
     [5] = ITEM5_FRIENDSHIP_ALL,
     [6] = -5,  // Friendship change, low
@@ -104,7 +104,7 @@ const u8 gItemEffect_HealPowder[9] = {
     [8] = -10, // Friendship change, high
 };
 
-const u8 gItemEffect_RevivalHerb[10] = {
+const u16 gItemEffect_RevivalHerb[10] = {
     [4] = ITEM4_REVIVE | ITEM4_HEAL_HP,
     [5] = ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_HEAL_HP_FULL,
@@ -113,48 +113,48 @@ const u8 gItemEffect_RevivalHerb[10] = {
     [9] = -20, // Friendship change, high
 };
 
-const u8 gItemEffect_Ether[7] = {
+const u16 gItemEffect_Ether[7] = {
     [4] = ITEM4_HEAL_PP_ONE | ITEM4_HEAL_PP,
     [6] = 10,
 };
 
-const u8 gItemEffect_MaxEther[7] = {
+const u16 gItemEffect_MaxEther[7] = {
     [4] = ITEM4_HEAL_PP_ONE | ITEM4_HEAL_PP,
     [6] = ITEM6_HEAL_PP_FULL,
 };
 
-const u8 gItemEffect_Elixir[7] = {
+const u16 gItemEffect_Elixir[7] = {
     [4] = ITEM4_HEAL_PP,
     [6] = 10, // Amount of PP to recover
 };
 
-const u8 gItemEffect_MaxElixir[7] = {
+const u16 gItemEffect_MaxElixir[7] = {
     [4] = ITEM4_HEAL_PP,
     [6] = ITEM6_HEAL_PP_FULL,
 };
 
-const u8 gItemEffect_LavaCookie[6] = {
+const u16 gItemEffect_LavaCookie[6] = {
     [3] = ITEM3_STATUS_ALL,
 };
 
-const u8 gItemEffect_BlueFlute[6] = {
+const u16 gItemEffect_BlueFlute[6] = {
     [3] = ITEM3_SLEEP,
 };
 
-const u8 gItemEffect_YellowFlute[6] = {
+const u16 gItemEffect_YellowFlute[6] = {
     [3] = ITEM3_CONFUSION,
 };
 
-const u8 gItemEffect_RedFlute[6] = {
+const u16 gItemEffect_RedFlute[6] = {
     [0] = ITEM0_INFATUATION,
 };
 
-const u8 gItemEffect_BerryJuice[7] = {
+const u16 gItemEffect_BerryJuice[7] = {
     [4] = ITEM4_HEAL_HP,
     [6] = 20, // Amount of HP to recover
 };
 
-const u8 gItemEffect_SacredAsh[7] = {
+const u16 gItemEffect_SacredAsh[7] = {
     [0] = ITEM0_SACRED_ASH,
     [4] = ITEM4_REVIVE | ITEM4_HEAL_HP,
     [6] = ITEM6_HEAL_HP_FULL,
@@ -165,39 +165,39 @@ const u8 gItemEffect_SacredAsh[7] = {
     [(i) + 1] = 3, /* Friendship change, mid */  \
     [(i) + 2] = 2  /* Friendship change, high */
 
-const u8 gItemEffect_HPUp[10] = {
+const u16 gItemEffect_HPUp[10] = {
     [4] = ITEM4_EV_HP,
     [5] = ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_ADD_EV,
     VITAMIN_FRIENDSHIP_CHANGE(7),
 };
 
-const u8 gItemEffect_Protein[10] = {
+const u16 gItemEffect_Protein[10] = {
     [4] = ITEM4_EV_ATK,
     [5] = ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_ADD_EV,
     VITAMIN_FRIENDSHIP_CHANGE(7),
 };
 
-const u8 gItemEffect_Iron[10] = {
+const u16 gItemEffect_Iron[10] = {
     [5] = ITEM5_EV_DEF | ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_ADD_EV,
     VITAMIN_FRIENDSHIP_CHANGE(7),
 };
 
-const u8 gItemEffect_Carbos[10] = {
+const u16 gItemEffect_Carbos[10] = {
     [5] = ITEM5_EV_SPEED | ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_ADD_EV,
     VITAMIN_FRIENDSHIP_CHANGE(7),
 };
 
-const u8 gItemEffect_Calcium[10] = {
+const u16 gItemEffect_Calcium[10] = {
     [5] = ITEM5_EV_SPATK | ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_ADD_EV,
     VITAMIN_FRIENDSHIP_CHANGE(7),
 };
 
-const u8 gItemEffect_RareCandy[10] = {
+const u16 gItemEffect_RareCandy[10] = {
     [3] = ITEM3_LEVEL_UP,
     [4] = ITEM4_REVIVE | ITEM4_HEAL_HP,
     [5] = ITEM5_FRIENDSHIP_ALL,
@@ -205,19 +205,19 @@ const u8 gItemEffect_RareCandy[10] = {
     VITAMIN_FRIENDSHIP_CHANGE(7),
 };
 
-const u8 gItemEffect_PPUp[9] = {
+const u16 gItemEffect_PPUp[9] = {
     [4] = ITEM4_PP_UP,
     [5] = ITEM5_FRIENDSHIP_ALL,
     VITAMIN_FRIENDSHIP_CHANGE(6),
 };
 
-const u8 gItemEffect_Zinc[10] = {
+const u16 gItemEffect_Zinc[10] = {
     [5] = ITEM5_EV_SPDEF | ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_ADD_EV,
     VITAMIN_FRIENDSHIP_CHANGE(7),
 };
 
-const u8 gItemEffect_PPMax[9] = {
+const u16 gItemEffect_PPMax[9] = {
     [5] = ITEM5_PP_MAX | ITEM5_FRIENDSHIP_ALL,
     VITAMIN_FRIENDSHIP_CHANGE(6),
 };
@@ -226,7 +226,7 @@ const u8 gItemEffect_PPMax[9] = {
     [6] = 1, /* Friendship change, low */ \
     [7] = 1  /* Friendship change, mid */
 
-const u8 gItemEffect_GuardSpec[8] = {
+const u16 gItemEffect_GuardSpec[8] = {
     [3] = ITEM3_GUARD_SPEC,
     [5] = ITEM5_FRIENDSHIP_LOW | ITEM5_FRIENDSHIP_MID,
     STAT_BOOST_FRIENDSHIP_CHANGE,
@@ -236,105 +236,105 @@ const u8 gItemEffect_GuardSpec[8] = {
 // only uses the least significant bit of its full mask.
 // The full constant is commented next to it
 
-const u8 gItemEffect_DireHit[8] = {
+const u16 gItemEffect_DireHit[8] = {
     [0] = 1 << 5, // ITEM0_DIRE_HIT
     [5] = ITEM5_FRIENDSHIP_LOW | ITEM5_FRIENDSHIP_MID,
     STAT_BOOST_FRIENDSHIP_CHANGE,
 };
 
-const u8 gItemEffect_XAttack[8] = {
+const u16 gItemEffect_XAttack[8] = {
     [0] = 1, // ITEM0_X_ATTACK
     [5] = ITEM5_FRIENDSHIP_LOW | ITEM5_FRIENDSHIP_MID,
     STAT_BOOST_FRIENDSHIP_CHANGE,
 };
 
-const u8 gItemEffect_XDefend[8] = {
+const u16 gItemEffect_XDefend[8] = {
     [1] = 1 << 4, /// ITEM1_X_DEFEND
     [5] = ITEM5_FRIENDSHIP_LOW | ITEM5_FRIENDSHIP_MID,
     STAT_BOOST_FRIENDSHIP_CHANGE,
 };
 
-const u8 gItemEffect_XSpeed[8] = {
+const u16 gItemEffect_XSpeed[8] = {
     [1] = 1, // ITEM1_X_SPEED
     [5] = ITEM5_FRIENDSHIP_LOW | ITEM5_FRIENDSHIP_MID,
     STAT_BOOST_FRIENDSHIP_CHANGE,
 };
 
-const u8 gItemEffect_XAccuracy[8] = {
+const u16 gItemEffect_XAccuracy[8] = {
     [2] = 1 << 4, // ITEM2_X_ACCURACY
     [5] = ITEM5_FRIENDSHIP_LOW | ITEM5_FRIENDSHIP_MID,
     STAT_BOOST_FRIENDSHIP_CHANGE,
 };
 
-const u8 gItemEffect_XSpecial[8] = {
+const u16 gItemEffect_XSpecial[8] = {
     [2] = 1, // ITEM2_X_SPATK
     [5] = ITEM5_FRIENDSHIP_LOW | ITEM5_FRIENDSHIP_MID,
     STAT_BOOST_FRIENDSHIP_CHANGE,
 };
 
-const u8 gItemEffect_SunStone[6] = {
+const u16 gItemEffect_SunStone[6] = {
     [4] = ITEM4_EVO_STONE,
 };
 
-const u8 gItemEffect_MoonStone[6] = {
+const u16 gItemEffect_MoonStone[6] = {
     [4] = ITEM4_EVO_STONE,
 };
 
-const u8 gItemEffect_FireStone[6] = {
+const u16 gItemEffect_FireStone[6] = {
     [4] = ITEM4_EVO_STONE,
 };
 
-const u8 gItemEffect_ThunderStone[6] = {
+const u16 gItemEffect_ThunderStone[6] = {
     [4] = ITEM4_EVO_STONE,
 };
 
-const u8 gItemEffect_WaterStone[6] = {
+const u16 gItemEffect_WaterStone[6] = {
     [4] = ITEM4_EVO_STONE,
 };
 
-const u8 gItemEffect_LeafStone[6] = {
+const u16 gItemEffect_LeafStone[6] = {
     [4] = ITEM4_EVO_STONE,
 };
 
-const u8 gItemEffect_CheriBerry[6] = {
+const u16 gItemEffect_CheriBerry[6] = {
     [3] = ITEM3_PARALYSIS,
 };
 
-const u8 gItemEffect_ChestoBerry[6] = {
+const u16 gItemEffect_ChestoBerry[6] = {
     [3] = ITEM3_SLEEP,
 };
 
-const u8 gItemEffect_PechaBerry[6] = {
+const u16 gItemEffect_PechaBerry[6] = {
     [3] = ITEM3_POISON,
 };
 
-const u8 gItemEffect_RawstBerry[6] = {
+const u16 gItemEffect_RawstBerry[6] = {
     [3] = ITEM3_BURN,
 };
 
-const u8 gItemEffect_AspearBerry[6] = {
+const u16 gItemEffect_AspearBerry[6] = {
     [3] = ITEM3_FREEZE,
 };
 
-const u8 gItemEffect_LeppaBerry[7] = {
+const u16 gItemEffect_LeppaBerry[7] = {
     [4] = ITEM4_HEAL_PP_ONE | ITEM4_HEAL_PP,
     [6] = 10, // Amount of PP to recover
 };
 
-const u8 gItemEffect_OranBerry[7] = {
+const u16 gItemEffect_OranBerry[7] = {
     [4] = ITEM4_HEAL_HP,
     [6] = 10, // Amount of HP to recover
 };
 
-const u8 gItemEffect_PersimBerry[6] = {
+const u16 gItemEffect_PersimBerry[6] = {
     [3] = ITEM3_CONFUSION,
 };
 
-const u8 gItemEffect_LumBerry[6] = {
+const u16 gItemEffect_LumBerry[6] = {
     [3] = ITEM3_STATUS_ALL,
 };
 
-const u8 gItemEffect_SitrusBerry[7] = {
+const u16 gItemEffect_SitrusBerry[7] = {
     [4] = ITEM4_HEAL_HP,
     [6] = 30, // Amount of HP to recover
 };
@@ -344,45 +344,45 @@ const u8 gItemEffect_SitrusBerry[7] = {
     [8] = 5,  /* Friendship change, mid */  \
     [9] = 2   /* Friendship change, high */
 
-const u8 gItemEffect_PomegBerry[10] = {
+const u16 gItemEffect_PomegBerry[10] = {
     [4] = ITEM4_EV_HP,
     [5] = ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_SUBTRACT_EV,
     EV_BERRY_FRIENDSHIP_CHANGE,
 };
 
-const u8 gItemEffect_KelpsyBerry[10] = {
+const u16 gItemEffect_KelpsyBerry[10] = {
     [4] = ITEM4_EV_ATK,
     [5] = ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_SUBTRACT_EV,
     EV_BERRY_FRIENDSHIP_CHANGE,
 };
 
-const u8 gItemEffect_QualotBerry[10] = {
+const u16 gItemEffect_QualotBerry[10] = {
     [5] = ITEM5_EV_DEF | ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_SUBTRACT_EV,
     EV_BERRY_FRIENDSHIP_CHANGE,
 };
 
-const u8 gItemEffect_HondewBerry[10] = {
+const u16 gItemEffect_HondewBerry[10] = {
     [5] = ITEM5_EV_SPATK | ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_SUBTRACT_EV,
     EV_BERRY_FRIENDSHIP_CHANGE,
 };
 
-const u8 gItemEffect_GrepaBerry[10] = {
+const u16 gItemEffect_GrepaBerry[10] = {
     [5] = ITEM5_EV_SPDEF | ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_SUBTRACT_EV,
     EV_BERRY_FRIENDSHIP_CHANGE,
 };
 
-const u8 gItemEffect_TamatoBerry[10] = {
+const u16 gItemEffect_TamatoBerry[10] = {
     [5] = ITEM5_EV_SPEED | ITEM5_FRIENDSHIP_ALL,
     [6] = ITEM6_SUBTRACT_EV,
     EV_BERRY_FRIENDSHIP_CHANGE,
 };
 
-const u8 *const gItemEffectTable[] =
+const u16 *const gItemEffectTable[] =
 {
     [ITEM_POTION - ITEM_POTION]        = gItemEffect_Potion,
     [ITEM_ANTIDOTE - ITEM_POTION]      = gItemEffect_Antidote,
